prompt-shield/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
httpx>=0.27.0
anthropic>=0.25.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
echo "groq>=0.9.0" >> requirements.txt
pip install groq